% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_CNV.R
\name{annotate_amplicons_exact}
\alias{annotate_amplicons_exact}
\title{Annotate amplicon rowData with gene information via exact string matching}
\usage{
annotate_amplicons_exact(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment object}
}
\value{
MultiAssayExperiment with updated amplicons rowData
}
\description{
Annotate amplicon rowData with gene information via exact string matching
}
