% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_multiomics.R
\name{auto_annotate_clusters}
\alias{auto_annotate_clusters}
\title{Auto-annotate Seurat clusters based on surface markers}
\usage{
auto_annotate_clusters(seurat_obj, assay = "RNA")
}
\arguments{
\item{seurat_obj}{A Seurat object with protein assay}

\item{assay}{Name of the protein assay}
}
\description{
Auto-annotate Seurat clusters based on surface markers
}
