% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_multiomics.R
\name{compute_population_genotype_distribution}
\alias{compute_population_genotype_distribution}
\title{Compute genotype distribution for a specific cell population}
\usage{
compute_population_genotype_distribution(
  mae_data,
  variant_ids,
  cell_barcodes,
  use_compass,
  seurat_cluster = NULL
)
}
\arguments{
\item{mae_data}{MultiAssayExperiment object}

\item{variant_ids}{Character vector of targeted variants}

\item{cell_barcodes}{Character vector of barcodes in the population}

\item{use_compass}{Logical to use imputed data}
}
\description{
Compute genotype distribution for a specific cell population
}
