% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct.R
\name{estimate_scaling_factor}
\alias{estimate_scaling_factor}
\title{Estimation d'un facteur d'échelle global pour runs "oversequencés"
Heuristique : la doc NSP évoque 'scaling_factor(reads, jitter)' et une contrainte
  max_zero_read_cells : fraction max de cellules avec 0 reads après scaling.
Ici, on propose une stratégie robuste & simple, cohérente avec la doc.}
\usage{
estimate_scaling_factor(total_reads, max_zero_read_cells = 0.05)
}
\description{
Estimation d'un facteur d'échelle global pour runs "oversequencés"
Heuristique : la doc NSP évoque 'scaling_factor(reads, jitter)' et une contrainte
  max_zero_read_cells : fraction max de cellules avec 0 reads après scaling.
Ici, on propose une stratégie robuste & simple, cohérente avec la doc.
}
