% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_multiomics.R
\name{extract_variant_genotypes}
\alias{extract_variant_genotypes}
\title{Extract variant genotypes from MultiAssayExperiment or SummarizedExperiment}
\usage{
extract_variant_genotypes(mae_data, variant_id, use_compass)
}
\arguments{
\item{mae_data}{MultiAssayExperiment or SummarizedExperiment object}

\item{variant_id}{Character string of the targeted variant}

\item{use_compass}{Logical to select imputed vs raw assay}
}
\value{
data.frame with Barcode and Variant_Genotype columns
}
\description{
Extract variant genotypes from MultiAssayExperiment or SummarizedExperiment
}
