% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_api.R
\name{filter_and_annotate_variants}
\alias{filter_and_annotate_variants}
\title{Filter variants and fetch annotations}
\usage{
filter_and_annotate_variants(
  obj,
  paths,
  min_dp = 10,
  min_gq = 30,
  vaf_ref = 5,
  vaf_hom = 95,
  vaf_het = 30,
  min_cell_pt = 10,
  min_mut_cell_pt = 10,
  batch_size = 300
)
}
\arguments{
\item{obj}{ScIGMA_object instance.}

\item{paths}{List of API paths for fetch_variants_batch_fields.}

\item{min_dp}{Integer. Minimum depth.}

\item{min_gq}{Integer. Minimum genotype quality.}

\item{vaf_ref}{Numeric. Maximum VAF for reference call.}

\item{vaf_hom}{Numeric. Minimum VAF for homozygous alternate call.}

\item{vaf_het}{Numeric. Maximum VAF for heterozygous call.}

\item{min_cell_pt}{Numeric. Minimum percentage of cells covering a variant.}

\item{min_mut_cell_pt}{Numeric. Minimum percentage of mutated cells.}

\item{batch_size}{Integer. Batch size for API calls.}
}
\value{
Filtered and annotated ScIGMA_object.
}
\description{
Filter variants and fetch annotations
}
