% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_DNA.R
\name{generate_clone_palette}
\alias{generate_clone_palette}
\title{Generate stable color palette for DNA clones}
\usage{
generate_clone_palette(clone_levels)
}
\arguments{
\item{clone_levels}{Character vector or factor of clone names}
}
\value{
Named character vector of hex colors
}
\description{
Generate stable color palette for DNA clones
}
