% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_CNV.R
\name{generate_cnv_heatmap_filtered}
\alias{generate_cnv_heatmap_filtered}
\title{Prepare and generate CNV Heatmap for a specific region}
\usage{
generate_cnv_heatmap_filtered(
  obj,
  features = NULL,
  projection_type = "Position"
)
}
\arguments{
\item{obj}{A ScIGMA object containing `ploidy.mtx` and `mae`.}

\item{features}{Selection vector for the X axis (e.g., c("chr1", "chr2") or c("TP53", "KRAS")). If NULL, displays all.}

\item{projection_type}{Chosen projection type ("Position" or "Genes").}
}
\value{
A plot object (usually from ComplexHeatmap generated by plot_cnv_heatmap).
}
\description{
Filters the ploidy matrix of a ScIGMA object on specific
chromosomes or genes, and calls plot_cnv_heatmap().
}
