% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_CNV.R
\name{generate_cnv_lineplot_filtered}
\alias{generate_cnv_lineplot_filtered}
\title{Generate CNV Lineplot (Filtered)}
\usage{
generate_cnv_lineplot_filtered(
  obj,
  features = NULL,
  projection_type = "Genes+amplicons",
  clone = NULL
)
}
\arguments{
\item{obj}{A ScIGMA_data object containing the ploidy matrix (obj$ploidy.mtx).}

\item{features}{Optional vector of chromosomes (e.g., "chr1") or genes (e.g., "TP53") to filter.}

\item{projection_type}{"Position" or "Genes+amplicons".}

\item{clone}{The name of the clone to display (default is the first available).}
}
\value{
A plot object (Plotly or standard depending on plot_cnv_genome).
}
\description{
API function to generate the Ploidy profile lineplot in the same 
way as in the Shiny module `mod_analysis_CNV`.
}
