% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct.R
\name{generate_dna_variant_heatmap}
\alias{generate_dna_variant_heatmap}
\title{Fonction to generate the dna_variant heatmap}
\usage{
generate_dna_variant_heatmap(
  obj,
  selected_variants,
  n_cluster = 6,
  min_prop_cluster = 0.01
)
}
\description{
Fonction to generate the dna_variant heatmap
}
