% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_multiomics.R
\name{generate_multi_barplot_clones_per_gate}
\alias{generate_multi_barplot_clones_per_gate}
\title{Generate Multiomics Barplot of DNA Clones vs Gating Subpopulations}
\usage{
generate_multi_barplot_clones_per_gate(obj)
}
\description{
Generate Multiomics Barplot of DNA Clones vs Gating Subpopulations
}
