% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_multiomics.R
\name{generate_multi_barplot_gating_genotype}
\alias{generate_multi_barplot_gating_genotype}
\title{Generate Multiomics Barplot of Gating Subpopulation Genotype vs Variants}
\usage{
generate_multi_barplot_gating_genotype(
  obj,
  variant_ids,
  use_compass = TRUE,
  target_gate = "All"
)
}
\description{
Generate Multiomics Barplot of Gating Subpopulation Genotype vs Variants
}
