% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_multiomics.R
\name{generate_multi_barplot_variant_genotype}
\alias{generate_multi_barplot_variant_genotype}
\title{Generate Multiomics Barplot of DNA Variants vs Protein Clusters}
\usage{
generate_multi_barplot_variant_genotype(obj, variant_ids, use_compass = TRUE)
}
\description{
Generate Multiomics Barplot of DNA Variants vs Protein Clusters
}
