% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_multiomics.R
\name{generate_multi_umap_dna_clones}
\alias{generate_multi_umap_dna_clones}
\title{Generate Multiomics UMAP with DNA Clones}
\usage{
generate_multi_umap_dna_clones(obj, use_compass = TRUE)
}
\description{
Generate Multiomics UMAP with DNA Clones
}
