% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_multiomics.R
\name{generate_multi_umap_dna_genotype}
\alias{generate_multi_umap_dna_genotype}
\title{Generate Multiomics UMAP with DNA Genotype}
\usage{
generate_multi_umap_dna_genotype(obj, variant_id, use_compass = TRUE)
}
\description{
Generate Multiomics UMAP with DNA Genotype
}
