% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_protein.R
\name{generate_protein_barplot}
\alias{generate_protein_barplot}
\title{Generate Protein Barplot}
\usage{
generate_protein_barplot(obj, title = "Protein Percentage Distribution")
}
\arguments{
\item{obj}{A ScIGMA_data object.}

\item{title}{The plot title.}
}
\value{
An interactive Plotly object.
}
\description{
API function to generate the protein distribution barplot outside the Shiny context.
}
