% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_protein.R
\name{generate_protein_ridgeplot}
\alias{generate_protein_ridgeplot}
\title{Generate Protein Abundance Ridge Plot}
\usage{
generate_protein_ridgeplot(obj)
}
\arguments{
\item{obj}{A ScIGMA_data object.}
}
\value{
An interactive Plotly object.
}
\description{
API function to generate the protein ridge plot outside the Shiny context.
}
