% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_protein.R
\name{generate_protein_umap_clustering}
\alias{generate_protein_umap_clustering}
\title{Generate Protein UMAP with Unsupervised Clustering}
\usage{
generate_protein_umap_clustering(obj)
}
\description{
Generate Protein UMAP with Unsupervised Clustering
}
