% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_protein.R
\name{generate_protein_umap_gating}
\alias{generate_protein_umap_gating}
\title{Generate Protein UMAP with Gating}
\usage{
generate_protein_umap_gating(obj)
}
\description{
Generate Protein UMAP with Gating
}
