% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_protein.R
\name{generate_protein_umap_markers}
\alias{generate_protein_umap_markers}
\title{Generate Protein UMAP with Markers}
\usage{
generate_protein_umap_markers(obj, markers)
}
\description{
Generate Protein UMAP with Markers
}
