% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-config.R
\name{get_config_paths}
\alias{get_config_paths}
\title{Get Configuration Paths}
\usage{
get_config_paths()
}
\value{
A list of paths specified in the configuration file.
}
\description{
Retrieves path configurations from the `config.yml` file. It searches for the
configuration file within the package context and falls back to a local `inst/config.yml`
if not found.
}
