% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_compass.R
\name{get_imputed_genotypes}
\alias{get_imputed_genotypes}
\title{Reconstruct imputed single-cell genotypes from COMPASS outputs}
\usage{
get_imputed_genotypes(prefix_out)
}
\arguments{
\item{prefix_out}{Character. Le chemin et préfixe utilisé dans run_compass_mcmc}
}
\value{
Une matrice [Cellules x Locus] de type Integer (0 = REF, 1 = HET, 2 = HOM_ALT)
}
\description{
Reconstruct imputed single-cell genotypes from COMPASS outputs
}
