% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct.R
\name{get_no_missing_dna_variant}
\alias{get_no_missing_dna_variant}
\title{Fonction to generate the dna_variant heatmap}
\usage{
get_no_missing_dna_variant(
  obj,
  selected_variants_df,
  n_cluster = 6,
  min_prop_cluster = 0.01
)
}
\description{
Fonction to generate the dna_variant heatmap
}
