% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analysis_multiOmics.R
\name{mod_analysis_multiomics_server}
\alias{mod_analysis_multiomics_server}
\title{analysis_multiomics Server Functions}
\usage{
mod_analysis_multiomics_server(id, ScIGMA_data)
}
\description{
analysis_multiomics Server Functions
}
