% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analysis_multiomics.R
\name{mod_analysis_multiomics_ui}
\alias{mod_analysis_multiomics_ui}
\title{analysis_multiomics UI Function}
\usage{
mod_analysis_multiomics_ui(id)
}
\description{
analysis_multiomics UI Function
}
