% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_protein.R
\name{normalize_protein_asinh}
\alias{normalize_protein_asinh}
\title{Normalize Protein Data using asinh}
\usage{
normalize_protein_asinh(obj, cofactor = 5)
}
\arguments{
\item{obj}{A `ScIGMA_object` (R6).}

\item{cofactor}{Numeric. The cofactor for asinh transformation.}
}
\value{
The input `ScIGMA_object` with `protein.mtx` normalized and
 `protein.normalize.method` set to "asinh normalized".
}
\description{
Normalizes protein expression data using asinh transformation.
}
