% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_protein.R
\name{normalize_protein_dsb}
\alias{normalize_protein_dsb}
\title{Normalize Protein Data using DSB (Denoised Scaled by Background)}
\usage{
normalize_protein_dsb(obj, min_bg_counts = 10, max_bg_counts = 500)
}
\arguments{
\item{obj}{A `ScIGMA_object` (R6).}

\item{min_bg_counts}{Minimum number of background counts to include an empty droplet.}

\item{max_bg_counts}{Maximum number of background counts to include an empty droplet.}
}
\value{
The input `ScIGMA_object` with `protein.mtx` normalized and
 `protein.normalize.method` set to "DSB normalized".
}
\description{
Normalizes protein expression data using the DSB method, which corrects for background
noise using empty droplets.
}
