% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_protein.R
\name{normalize_protein_nsp}
\alias{normalize_protein_nsp}
\title{Normalize Protein Data using NSP}
\usage{
normalize_protein_nsp(obj)
}
\arguments{
\item{obj}{A `ScIGMA_object` (R6).}
}
\value{
The input `ScIGMA_object` with `protein.mtx` normalized and
 `protein.normalize.method` set to "NSP normalized".
}
\description{
Normalizes protein expression data using the NSP method.
}
