% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_TapestriNSP.R
\name{normalize_reads}
\alias{normalize_reads}
\title{Normalize Protein Reads}
\usage{
normalize_reads(
  reads,
  method = "CLR",
  jitter = 0.5,
  scale = NULL,
  sample_size = 1000,
  random_state = NULL
)
}
\arguments{
\item{reads}{A numeric matrix of raw reads (rows = cells, columns = antibodies).}

\item{method}{Character string: "CLR", "asinh", "NSP" or "ANSP".}

\item{jitter}{Standard deviation of the Gaussian noise to add (default = 0.5). Useful for "NSP" and "asinh".}

\item{scale}{Scaling factor (for "NSP" / "ANSP").}

\item{sample_size}{Number of cells to use to estimate model parameters ("ANSP").}

\item{random_state}{Integer to set the random seed (reproducibility).}
}
\value{
A list containing the `normalized_counts` matrix and the `nsp_model` object (if applicable).
}
\description{
Wrapper function to normalize Tapestri protein read counts.
}
