% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct.R
\name{normalizeProtein}
\alias{normalizeProtein}
\title{Protein matrix normalization}
\usage{
normalizeProtein(ScIGMA_object)
}
\arguments{
\item{ScIGMA_object}{A `ScIGMA_object` (R6).}
}
\value{
The input `ScIGMA_object` with `protein.mtx` normalized and
 `protein.normalize.method` set to "normalized".
}
\description{
Normalizes the protein matrix within a `ScIGMA_object` using the
centered log‑ratio (CLR) transform, inspired by the analogous function
in the optima package.
}
\examples{
\dontrun{
scigma <- normalizeProtein(scigma)
}
}
\keyword{ScIGMA}
\keyword{normalization}
\keyword{protein}
