% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_CNV.R
\name{plot_cnv_genome}
\alias{plot_cnv_genome}
\title{Plot Genomic CNV Profile (Ploidy) - Bold & Larger Labels}
\usage{
plot_cnv_genome(
  cnv_matrix,
  gene_annotation = NULL,
  sub_indices = NULL,
  title = "Genomic Ploidy Profile",
  max_points = 500,
  lineplot_type = "Genes+amplicons"
)
}
\description{
Plot Genomic CNV Profile (Ploidy) - Bold & Larger Labels
}
