% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_CNV.R
\name{plot_cnv_heatmap}
\alias{plot_cnv_heatmap}
\title{Plot CNV Heatmap}
\usage{
plot_cnv_heatmap(obj, ploidy_data, display_gene = FALSE)
}
\arguments{
\item{obj}{A ScIGMA object containing a `cnv_id_table` with `dna_id`, `chrom`, and `start_pos` columns.}

\item{ploidy_data}{A matrix or data frame of ploidy values (result from `process_cnv_to_clonal_profile`).}
}
\value{
A `Heatmap` object (from the ComplexHeatmap package).
}
\description{
Generates a heatmap of Copy Number Variation (CNV) ploidy data.
The function aligns the provided ploidy data with chromosomal positions defined
in the ScIGMA object and visualizes it using `ComplexHeatmap`.
}
