% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_compass.R
\name{prepare_compass_from_csv}
\alias{prepare_compass_from_csv}
\title{Parse COMPASS CSV outputs into memory-safe matrices for Rcpp}
\usage{
prepare_compass_from_csv(variants_file, regions_file)
}
\arguments{
\item{variants_file}{Path to the [sample]_variants.csv file}

\item{regions_file}{Path to the [sample]_regions.csv file}
}
\value{
A list formatted strictly for run_compass_mcmc() API
}
\description{
Parse COMPASS CSV outputs into memory-safe matrices for Rcpp
}
