% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_CNV.R
\name{process_cnv_to_clonal_profile}
\alias{process_cnv_to_clonal_profile}
\title{Process CNV Matrix to Diploid-Referenced Clonal Profiles}
\usage{
process_cnv_to_clonal_profile(
  filtered_data,
  dna_variant_clones,
  diploid_ref = "2",
  exclude_clone = "small"
)
}
\arguments{
\item{filtered_data}{A list containing the CNV matrix under
`$filtered_cnv_mtx` (amplicons x cells).}

\item{dna_variant_clones}{A named vector or factor where names are cell barcodes
and values are clone IDs.}

\item{diploid_ref}{The ID (string) of the clone group to be used as the 2N reference.}

\item{exclude_clone}{A character string of a clone to exclude from the final profile
(e.g., "small" or NA).}
}
\value{
A matrix where rows are genomic regions (amplicons) and columns are
  clonal profiles, normalized to the diploid reference.
}
\description{
Normalizes read counts, aggregates the normalized data by clone
  (using median), and calculates the final profile relative to a specified
  diploid reference clone.
}
