% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct.R
\name{render_protein_ridge_plot}
\alias{render_protein_ridge_plot}
\title{Render ridge plot with ggplot2 and plotly}
\usage{
render_protein_ridge_plot(obj)
}
\description{
Render ridge plot with ggplot2 and plotly
}
