% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct.R
\name{sanitize_gate_name}
\alias{sanitize_gate_name}
\title{Sanitize manual gating names to ensure strictly valid R strings}
\usage{
sanitize_gate_name(gate_name)
}
\arguments{
\item{gate_name}{Character vector of raw gate names}
}
\value{
Character vector of sanitized names
}
\description{
Sanitize manual gating names to ensure strictly valid R strings
}
