% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_H5Processing.R
\name{sanitize_mae_strings}
\alias{sanitize_mae_strings}
\title{Sanitize MAE string metadata to remove Windows carriage returns and whitespaces}
\usage{
sanitize_mae_strings(mae)
}
\arguments{
\item{mae}{MultiAssayExperiment object}
}
\value{
Sanitized MultiAssayExperiment
}
\description{
Sanitize MAE string metadata to remove Windows carriage returns and whitespaces
}
