% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScIGMA-package.R
\docType{package}
\name{ScIGMA-package}
\alias{ScIGMA}
\alias{ScIGMA-package}
\title{ScIGMA: ScIGMA – Single-cell Integrated Genomic & Multi-omics Analyzer}
\description{
PKG_DESC.
}
\author{
\strong{Maintainer}: Geoffrey Pawlak \email{pawlak.geoffrey@gmail.com}

}
\keyword{internal}
