% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ScIGMA_R6.R
\name{ScIGMA_object}
\alias{ScIGMA_object}
\title{ScIGMA Object Class}
\description{
This class encapsulates various matrices and metadata required for single-cell genomics and proteomics analysis.
}
\details{
An R6 class for managing and analyzing single-cell genomic and proteomic data within a Shiny application built using golem.


Contains methods for initializing, summarizing, and manipulating single-cell variant, amplification, and protein data.
}
\examples{
scigma <- ScIGMA_object$new(meta.data = c("Sample1", "Sample2"), cell.ids = c("Cell1", "Cell2"))
scigma$add_variant("Variant1")
scigma$summary()

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{meta.data}}{Metadata information}

\item{\code{cell.ids}}{Cell identifiers}

\item{\code{cell.labels}}{Cell labels}

\item{\code{variants}}{Variant information}

\item{\code{variant.filter}}{Filtered variant information}

\item{\code{vaf.mtx}}{Variant allele frequency matrix}

\item{\code{vaf.mtx.cells}}{Cells associated with VAF matrix}

\item{\code{gt.mtx}}{Genotype matrix}

\item{\code{gt.mtx.cells}}{Cells associated with genotype matrix}

\item{\code{dp.mtx}}{Depth matrix}

\item{\code{dp.mtx.cells}}{Cells associated with depth matrix}

\item{\code{gq.mtx}}{Genotype quality matrix}

\item{\code{gq.mtx.cells}}{Cells associated with genotype quality matrix}

\item{\code{amps}}{Amplification information}

\item{\code{amp.normalize.method}}{Amplification normalization method}

\item{\code{amp.mtx}}{Amplification matrix}

\item{\code{amp.mtx.cells}}{Cells associated with amplification matrix}

\item{\code{ploidy.mtx}}{Ploidy matrix}

\item{\code{proteins}}{Protein information}

\item{\code{protein.normalize.method}}{Protein normalization method}

\item{\code{protein.mtx}}{Protein expression matrix}

\item{\code{protein.mtx.cells}}{Cells associated with protein matrix}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ScIGMA_object-new}{\code{ScIGMA_object$new()}}
\item \href{#method-ScIGMA_object-print}{\code{ScIGMA_object$print()}}
\item \href{#method-ScIGMA_object-summary}{\code{ScIGMA_object$summary()}}
\item \href{#method-ScIGMA_object-add_variant}{\code{ScIGMA_object$add_variant()}}
\item \href{#method-ScIGMA_object-filter_variants}{\code{ScIGMA_object$filter_variants()}}
\item \href{#method-ScIGMA_object-clone}{\code{ScIGMA_object$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScIGMA_object-new"></a>}}
\if{latex}{\out{\hypertarget{method-ScIGMA_object-new}{}}}
\subsection{Method \code{new()}}{
Constructor for the ScIGMA object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScIGMA_object$new(
  meta.data = character(),
  cell.ids = character(),
  cell.labels = character(),
  variants = character(),
  variant.filter = character(),
  vaf.mtx = matrix(),
  vaf.mtx.cells = character(),
  gt.mtx = matrix(),
  gt.mtx.cells = character(),
  dp.mtx = matrix(),
  dp.mtx.cells = character(),
  gq.mtx = matrix(),
  gq.mtx.cells = character(),
  amps = character(),
  amp.normalize.method = character(),
  amp.mtx = matrix(),
  amp.mtx.cells = character(),
  ploidy.mtx = matrix(),
  proteins = character(),
  protein.normalize.method = character(),
  protein.mtx = matrix(),
  protein.mtx.cells = character()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{meta.data}}{Metadata for samples}

\item{\code{cell.ids}}{Cell identifiers}

\item{\code{cell.labels}}{Cell labels}

\item{\code{variants}}{Variant information}

\item{\code{variant.filter}}{Filtered variants}

\item{\code{vaf.mtx}}{Variant allele frequency matrix}

\item{\code{vaf.mtx.cells}}{Cells associated with VAF matrix}

\item{\code{gt.mtx}}{Genotype matrix}

\item{\code{gt.mtx.cells}}{Cells associated with genotype matrix}

\item{\code{dp.mtx}}{Depth matrix}

\item{\code{dp.mtx.cells}}{Cells associated with depth matrix}

\item{\code{gq.mtx}}{Genotype quality matrix}

\item{\code{gq.mtx.cells}}{Cells associated with genotype quality matrix}

\item{\code{amps}}{Amplification data}

\item{\code{amp.normalize.method}}{Method for normalizing amplification data}

\item{\code{amp.mtx}}{Amplification matrix}

\item{\code{amp.mtx.cells}}{Cells associated with amplification matrix}

\item{\code{ploidy.mtx}}{Ploidy matrix}

\item{\code{proteins}}{Protein data}

\item{\code{protein.normalize.method}}{Method for normalizing protein data}

\item{\code{protein.mtx}}{Protein matrix}

\item{\code{protein.mtx.cells}}{Cells associated with protein matrix}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScIGMA_object-print"></a>}}
\if{latex}{\out{\hypertarget{method-ScIGMA_object-print}{}}}
\subsection{Method \code{print()}}{
Print method for displaying a summary of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScIGMA_object$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScIGMA_object-summary"></a>}}
\if{latex}{\out{\hypertarget{method-ScIGMA_object-summary}{}}}
\subsection{Method \code{summary()}}{
Provides a concise summary of key attributes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScIGMA_object$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScIGMA_object-add_variant"></a>}}
\if{latex}{\out{\hypertarget{method-ScIGMA_object-add_variant}{}}}
\subsection{Method \code{add_variant()}}{
Add a variant to the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScIGMA_object$add_variant(variant)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{variant}}{Variant to be added}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScIGMA_object-filter_variants"></a>}}
\if{latex}{\out{\hypertarget{method-ScIGMA_object-filter_variants}{}}}
\subsection{Method \code{filter_variants()}}{
Filters variants based on a provided function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScIGMA_object$filter_variants(filter_fun)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filter_fun}}{Filtering function to apply}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScIGMA_object-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ScIGMA_object-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScIGMA_object$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
