% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct.R
\name{select_bg_sig_indices}
\alias{select_bg_sig_indices}
\title{Sélectionne indices "background" (bas) et "signal" (haut) pour un vecteur y}
\usage{
select_bg_sig_indices(y, p_low = 0.1, p_high = 0.9, min_points = 20L)
}
\description{
Sélectionne indices "background" (bas) et "signal" (haut) pour un vecteur y
}
