% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct.R
\name{update_cluster_labels}
\alias{update_cluster_labels}
\title{Update cluster labels}
\usage{
update_cluster_labels(new_labels)
}
\arguments{
\item{new_labels}{A named vector where names are old levels and values are new labels.}
}
\description{
Update cluster labels safely.
}
