% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ScIGMA_R6.R
\name{ScIGMA_object}
\alias{ScIGMA_object}
\title{#' @importFrom R6 R6Class
#' @importFrom DelayedArray DelayedArray realize
#' @importFrom HDF5Array HDF5Array writeHDF5Array setHDF5DumpDir setHDF5DumpFile
#' @keywords classes
ScIGMA_object <- R6::R6Class(
    classname = "ScIGMA_object",
    public = list(
        #' @field data Named list holding all payload fields
        #' (matrices/vectors/metadata).
        data = NULL,}
\description{

}
\details{

}
\keyword{classes}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ScIGMA_object-new}{\code{ScIGMA_object$new()}}
\item \href{#method-ScIGMA_object-print}{\code{ScIGMA_object$print()}}
\item \href{#method-ScIGMA_object-realize_all}{\code{ScIGMA_object$realize_all()}}
\item \href{#method-ScIGMA_object-clone}{\code{ScIGMA_object$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScIGMA_object-new"></a>}}
\if{latex}{\out{\hypertarget{method-ScIGMA_object-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScIGMA_object$new(
  mae = NULL,
  mae_raw = NULL,
  backing_files = list(),
  filetype = NULL,
  seurat_object = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScIGMA_object-print"></a>}}
\if{latex}{\out{\hypertarget{method-ScIGMA_object-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScIGMA_object$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScIGMA_object-realize_all"></a>}}
\if{latex}{\out{\hypertarget{method-ScIGMA_object-realize_all}{}}}
\subsection{Method \code{realize_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScIGMA_object$realize_all(
  dir,
  file = "scigma_store.h5",
  chunkdim = NULL,
  level = 6
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScIGMA_object-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ScIGMA_object-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScIGMA_object$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
